% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp2.R
\name{dp2}
\alias{dp2}
\title{DP2 Function}
\usage{
dp2(x, iterations, epsilon = 1e-04, polarity = NULL, qualitative = NULL)
}
\arguments{
\item{x}{A numeric matrix or data frame which provides the data set with all the study variables.}

\item{iterations}{maximum number of iterations for the algorithm}

\item{epsilon}{A double maximum number of distance accepted.}

\item{polarity}{polarity}

\item{qualitative}{numeric vector which provides the position of qualitative variables}
}
\value{
Return the a matrix with the variance inflation factor and a boolean variable if there is multicollinearity
}
\description{
This function computes a indicator
}
\examples{

data(NUTS2)
C <- NUTS2[,4:11]
polarity <- c(5,6,7,8)
dp2(C, iterations=10 ,polarity = polarity)

}
\references{
Jiménez-Fernández, E., Sánchez, A., & Ortega-Pérez, M. (2022). Dealing with weighting scheme in composite indicators: an unsupervised distance-machine learning proposal for quantitative data. Socio-Economic Planning Sciences, 83, 101339.
}
\seealso{
Fabio
}
\author{
Fabio Sepulveda:  \email{fhsepulveda@udemedellin.edu.co}
Andrés Vinueza:   \email{andres.vinueza@epn.edu.ec}
}
