% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_analysis.R
\name{index_analysis}
\alias{index_analysis}
\title{Perform INtron differences to EXons (INdEX) analysis}
\usage{
index_analysis(exon, intron, group, design = NULL, contrast = NULL,
  p.value = 0.05)
}
\arguments{
\item{exon}{the DGE list containing exon information}

\item{intron}{the DGE list containing intron information}

\item{group}{the vector of groups}

\item{design}{the design matrix}

\item{contrast}{the contrast to test}

\item{p.value}{the p-value for decideTests}
}
\value{
a list containing a combined decideTest result, the input dges, voom objects and tops for exon and introns.
}
\description{
Perform INtron differences to EXons (INdEX) analysis
}
