% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_normalization.R
\name{matrix_normalization}
\alias{matrix_normalization}
\title{Normalization}
\usage{
matrix_normalization(x, polarity = NULL)
}
\arguments{
\item{x}{A numeric matrix or data frame which provides the data set with all the study variables.}

\item{polarity}{A numeric vector which indicates the relation (polarity) among the latent phenomenon and all the study variables.By default is null, if the relation is negative.}
}
\value{
<<<<<<< HEAD
Return the normalizex matrix. nnn
=======
Return the normalizex matriz.
>>>>>>> 1e21aaff21f3a3316a73860b370b3896628444d1
}
\description{
This function Computes the normalized matrix.
}
\references{
Jiménez-Fernández, E., Sánchez, A., & Ortega-Pérez, M. (2022). Dealing with weighting scheme in composite indicators: an unsupervised distance-machine learning proposal for quantitative data. Socio-Economic Planning Sciences, 83, 101339.
}
\seealso{
Fabio
}
\author{
Fabio Sepulveda:  \email{fhsepulveda@udemedellin.edu.co}
}
