% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicollinearity.R
\name{multicollinearity}
\alias{multicollinearity}
\title{Multicollinearity}
\usage{
multicollinearity(x)
}
\arguments{
\item{x}{A numeric matrix or data frame which provides the data set with all the study variables.}
}
\value{
Return the a matrix with the variance inflation factor and a boolean variable if there is multicollinearity
}
\description{
This function computes a matrix with the variance inflation factor and evaluate if the variance inflation factor is greater than 10
}
\examples{

data(NUTS2)
B <- NUTS2[,4:11]
multicollinearity(B)

}
\references{
Jiménez-Fernández, E., Sánchez, A., & Ortega-Pérez, M. (2022). Dealing with weighting scheme in composite indicators: an unsupervised distance-machine learning proposal for quantitative data. Socio-Economic Planning Sciences, 83, 101339.
}
\seealso{
Fabio
}
\author{
Fabio Sepulveda:  \email{fhsepulveda@udemedellin.edu.co}
Andrés Vinueza:   \email{andres.vinueza@epn.edu.ec}
}
