% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lcpm_cor.R
\name{plot_lcpm_cor}
\alias{plot_lcpm_cor}
\title{Plot correlation and bar plots of logCPM between introns and exons}
\usage{
plot_lcpm_cor(index_output)
}
\arguments{
\item{index_output}{the output from running index_analysis}
}
\value{
None
}
\description{
Plot correlation and bar plots of logCPM between introns and exons
}
