% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_voom.R
\name{plot_voom}
\alias{plot_voom}
\title{Plot voom for introns and exons}
\usage{
plot_voom(index_output)
}
\arguments{
\item{index_output}{the output from running index_analysis}
}
\value{
None
}
\description{
Plot voom for introns and exons
}
