% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{Quantification}
\usage{
quantify(x, y)
}
\arguments{
\item{x}{A data frame which provides the data set with all the study variables.}

\item{y}{A numeric vector which indicates the position of the columns of type factor.}
}
\value{
Return a data frame with quantified data.
}
\description{
This function quantifies qualitative variable.
}
\examples{

data(LondonHouse)
LondonHouse$BATH2 = factor(LondonHouse$BATH2)
LondonHouse$GARAGE1 = factor(LondonHouse$GARAGE1)
LondonHouse$BEDS2 = factor(LondonHouse$BEDS2)
qualitative = c(3,4,5)
quantify(LondonHouse,qualitative)

}
\references{
Gifi, A. (1990). Nonlinear Multivariate Analysis. New York: Wiley
}
\seealso{
Gifi
}
\author{
Eduardo Jiménez: \email{edjimfer@ugr.es}
}
