\name{INdEX-package}
\alias{INdEX-package}
\alias{INdEX}
\docType{package}
\title{
\packageTitle{INdEX}
}
\description{
\packageDescription{INdEX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{INdEX}
\packageIndices{INdEX}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageMaintainer{INdEX}
}
\references{

}

\keyword{ package }
\examples{
~~ simple examples of the most important functions ~~
}
