% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Create main INdEX analysis plots}
\usage{
plot_index(index_output, type = c("tstat", "logfc"))
}
\arguments{
\item{index_output}{the output from running index_analysis}
}
\value{
None
}
\description{
Create main INdEX analysis plots
}
