% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{Quantification}
\usage{
quantify(x, y)
}
\arguments{
\item{x}{A data frame which provides the data set with all the study variables.}

\item{y}{A numeric vector which indicates the position of the columns of type factor.}
}
\value{
Return a data frame with quantified data.
}
\description{
This function quantifies qualitative variables.
}
\references{
Gifi, A. (1990). Nonlinear Multivariate Analysis. New York: Wiley
}
\seealso{
Fabio
}
\author{
Fabio Sepulveda:  \email{fhsepulveda@udemedellin.edu.co}
}
